Fixed errors with null values present
Can now serialize jsonnode.Node values. Fields are sorted in a "smart"
way utilizing a topological sort. Data frames now only return
json.RawMessage values or time.Time values. This (currently) has the
downside of allowing string values to be interpreted numerically if the
contents of a string are numeric.