Frame exclude feature
This feature makes it easy to exclude any label from the data frame.
This is necessary because when numbers are included in the data frame, they are always used in the graph.
With this feature you can exclude any label from the data frame, allowing numbers to be used as labels.